Move painter functions into paint.c. Fix mistake in paint_attach_to_previous_ps
