+++ openrct2.vcxproj
@@ -58,6 +58,7 @@
     <ClCompile Include="src\interface\chat.c" />
     <ClCompile Include="src\interface\colour.c" />
     <ClCompile Include="src\interface\paint_surface.c" />
+    <ClCompile Include="src\interface\paint.c" />
     <ClCompile Include="src\interface\Theme.cpp" />
     <ClCompile Include="src\interface\console.c" />
     <ClCompile Include="src\interface\graph.c" />
@@ -241,6 +242,7 @@
     <ClInclude Include="src\interface\chat.h" />
     <ClInclude Include="src\interface\colour.h" />
     <ClInclude Include="src\interface\paint_surface.h" />
+    <ClInclude Include="src\interface\paint.h" />
     <ClInclude Include="src\interface\themes.h" />
     <ClInclude Include="src\interface\console.h" />
     <ClInclude Include="src\interface\graph.h" />
