+++ openrct2.vcxproj.filters
@@ -604,6 +604,7 @@
       <Filter>Source\CommandLine</Filter>
     </ClCompile>
     <ClCompile Include="src\interface\paint_surface.c" />
+    <ClCompile Include="src\interface\paint.c" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\management\award.h">
@@ -925,6 +926,7 @@
       <Filter>Source\RCT1</Filter>
     </ClInclude>
     <ClInclude Include="src\interface\paint_surface.h" />
+    <ClInclude Include="src\interface\paint.h" />
   </ItemGroup>
   <ItemGroup>
     <Image Include="resources\logo\icon.ico">
